pyproject.toml
[tool.poetry]
name = "data-analyser-backend"
version = "0.1.0"
description = "Backend of Hack the Insurance Competition project"
authors = ["Eray Aynacı <73905858+er4y-c@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
dynaconf = "^3.2.4"
psycopg2-binary = "^2.9.9"
pandas = "^2.1.4"
python-multipart = "^0.0.6"
pymongo = "^4.6.1"
passlib = "^1.7.4"
jwt = "^1.3.1"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
scikit-learn = "^1.3.2"
xgboost = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
